Refactor connection handling and token generation logic
Moved connection handling functions to a new dedicated `conn.go` file for better modularization and clarity. Simplified token TTL handling by replacing redundant logic with a single default and updated schema types. This improves code maintainability and reduces duplication.

+++ pkg/api/schema.go
@@ -2,11 +2,11 @@ package api
 
 type GenerateTokenRequest struct {
 	KeyID string `json:"key_id"`
-	TTL   int64  `json:"ttl"`
+	TTL   int    `json:"ttl"`
 }
 
 type GenerateTokenResponse struct {
 	Token string `json:"token"`
 	KeyID string `json:"key_id"`
-	TTL   int64  `json:"ttl"`
+	TTL   int    `json:"ttl"`
 }
